Fix resetting visits state on app opening (#23)
The visits update request was triggered by opening the app before, which
caused the state to reset to the default date range.
It's fixed by limiting the trigger to only the post-login refresh
action.

Also the reducer updated the list data with the wrong dates, so the
range view and list data were out of sync.
It's fixed by rejecting the action if it updates the view for the wrong
date range (the correct update first updates the state for it to match
the requested dates)